Refactor: centralize test database and data insertion helpers into `testUtils`
